// generated/prisma/enums.ts
export const users_role = {
  admin: 'admin',
  user: 'user',
  manager: 'manager',
  operator: 'operator',
  sales: 'sales',
} as const;

export type users_role = (typeof users_role)[keyof typeof users_role];